make counters match input argument size
Since the intput sizes are size_t, make the internal counter
variables use the same size.  The internal counters might
have been a larger type and this extra space would have
never been used since it is limited by the input size.

+++ sort/c/sort.c
@@ -4,7 +4,7 @@
 #if BUBBLE_SORT
 void sort(char **lines, size_t size)
 {
-	unsigned int i;
+	size_t i;
 	char change;
 	char *tmp;
 
@@ -23,7 +23,7 @@ void sort(char **lines, size_t size)
 #elif INSERTION_SORT
 void sort(char **lines, size_t size)
 {
-	unsigned int i, j;
+	size_t i, j;
 
 	for (i = 1; i < size; i++) {
 		char *key = lines[i];
